update: 更新文件 wifi_event_subscriber_manager.cpp

+++ wifi/services/wifi_standard/wifi_framework/wifi_manage/wifi_sub_manage/wifi_event_subscriber_manager.cpp
@@ -1171,7 +1171,7 @@ void NetmgrEventSubscriber::OnReceiveEvent(const OHOS::EventFwk::CommonEventData
         WIFI_LOGI("NetmgrEventSubscriber OnReceiveEvent by AccTask %{public}d", accTaskState);
         WifiNetworkControlInfo networkControlInfo;
         networkControlInfo.sceneId = BG_LIMIT_CONTROL_ID_LOW_LATENCY;
-        networkControlInfo.state = static_cast<int>(accTaskState);
+        networkControlInfo.state = accTaskState;
         AppNetworkSpeedLimitService::GetInstance().ReceiveNetworkControlInfo(networkControlInfo);
         return;
     }
